move functions to tools
